Move `formatError` helper outside `src` directory.
